fix: change exception when tecnico have OS

+++ src/main/java/com/lfmelo/services/TecnicoService.java
@@ -48,6 +48,11 @@ public class TecnicoService {
 
 	public void delete(Integer id) {
 		Tecnico tecnico = this.findById(id);
+		
+		if (tecnico.getChamados().size() > 0) {
+			throw new DataIntegrityViolationException("Técnico possui ordens de serviço e não pode ser deletado!");
+		}
+		
 		this.repository.delete(tecnico);
 	}
